NGU_sistem/requirements.txt
# Основні залежності
Flask==2.0.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
joblib==1.1.0

# Для дашборду
dash==2.0.0
dash-bootstrap-components==1.0.0
plotly==5.3.1

# Для API
flask-restful==0.3.9
flask-cors==3.0.10

# Для роботи з даними
openpyxl==3.0.9
xlrd==2.0.1

# Для ML моделей
tensorflow==2.7.0
keras==2.7.0

# Утиліти
python-dotenv==0.19.1
tqdm==4.62.3
requests==2.26.0